Add max_length constraint to LineInput component
Add an optional max_length field to LineInputState that limits the
number of characters (not bytes) a user can type. When set, Insert
is rejected if the effective length would exceed the limit, Paste
content is truncated to fit, and SetValue is truncated to the max.
Selection replacements correctly account for characters being removed.

Also extract handle_event tests to a separate module to keep all
test files under the 1000-line limit.
